+++ blocklets/ai-studio/api/src/store/migrations/20240521162701-add-agent-input-secret.ts
@@ -1,0 +1,53 @@
+import { DataTypes } from 'sequelize';
+
+import type { Migration } from '../migrate';
+
+export const up: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.createTable('AgentInputSecrets', {
+    id: {
+      type: DataTypes.STRING,
+      primaryKey: true,
+      allowNull: false,
+    },
+    projectId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    targetProjectId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    targetAgentId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    targetInputKey: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    secret: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+    },
+    createdBy: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    updatedBy: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  });
+};
+
+export const down: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.dropTable('AgentInputSecrets');
+};