+++ packages/ai-runtime/src/core/assistant/index.ts
@@ -10,8 +10,8 @@ import { call, call as callFunc } from '@blocklet/sdk/lib/component';
 import { logger } from '@blocklet/sdk/lib/config';
 import axios, { isAxiosError } from 'axios';
 import { flattenDeep, isNil, pick, startCase, toLower } from 'lodash';
-import fetch from 'node-fetch';
 import { Worker } from 'snowflake-uuid';
+import { joinURL, withQuery } from 'ufo';
 import { NodeVM } from 'vm2';
 
 import { TranspileTs } from '../../builtin/complete';
@@ -78,6 +78,7 @@ const taskIdGenerator = new Worker();
 export const nextTaskId = () => taskIdGenerator.nextId().toString();
 
 export async function runAssistant({
+  getSecret,
   taskId,
   callAI,
   callAIImage,
@@ -92,6 +93,11 @@ export async function runAssistant({
   datastoreVariables,
   functionName,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   taskId: string;
   callAI: CallAI;
   callAIImage: CallAIImage;
@@ -102,7 +108,7 @@ export async function runAssistant({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
   functionName?: string;
 }): Promise<any> {
@@ -119,6 +125,7 @@ export async function runAssistant({
   });
 
   const assistantVariables = await getVariables({
+    getSecret,
     assistant,
     callAI,
     callAIImage,
@@ -152,6 +159,7 @@ export async function runAssistant({
 
     if (isPromptAssistant(assistant)) {
       return await runPromptAssistant({
+        getSecret,
         taskId,
         parentTaskId,
         callAI,
@@ -170,6 +178,7 @@ export async function runAssistant({
 
     if (isImageAssistant(assistant)) {
       return await runImageAssistant({
+        getSecret,
         taskId,
         parentTaskId,
         callAI,
@@ -221,6 +230,7 @@ export async function runAssistant({
 
     if (isRouterAssistant(assistant)) {
       return await runRouterAssistant({
+        getSecret,
         getAssistant,
         callAI,
         callAIImage,
@@ -318,9 +328,11 @@ export default async function(args) {
           return result;
         },
       },
+      fetch,
       URL,
       call,
-      fetch,
+      joinURL,
+      withQuery,
       ...args,
     },
   });
@@ -408,7 +420,7 @@ async function runApiAssistant({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
 }) {
   if (!assistant.requestUrl) throw new Error(`Assistant ${assistant.id}'s url is empty`);
@@ -487,6 +499,7 @@ async function runApiAssistant({
 const cacheTranslateFunctionNames: { [key: string]: string } = {};
 
 async function runRouterAssistant({
+  getSecret,
   callAI,
   callAIImage,
   taskId,
@@ -500,6 +513,11 @@ async function runRouterAssistant({
   projectId,
   datastoreVariables,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   callAI: CallAI;
   callAIImage: CallAIImage;
   taskId: string;
@@ -510,7 +528,7 @@ async function runRouterAssistant({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
 }) {
   if (!assistant.prompt) {
@@ -523,7 +541,7 @@ async function runRouterAssistant({
   const toolAssistants = (
     await Promise.all(
       routes.map(async (tool) => {
-        const toolAssistant = await getAssistant(tool.id);
+        const toolAssistant = await getAssistant(tool.id, { blockletDid: tool.blockletDid, projectId: tool.projectId });
         if (!toolAssistant) return undefined;
         const toolParameters = (toolAssistant.parameters ?? [])
           .filter((i): i is typeof i & Required<Pick<typeof i, 'key'>> => !!i.key && !tool.parameters?.[i.key])
@@ -845,6 +863,7 @@ async function runRouterAssistant({
           });
 
         const res = await runAssistant({
+          getSecret,
           taskId: currentTaskId,
           callAI,
           callAIImage,
@@ -1035,6 +1054,7 @@ const runRequestHistory = async ({
 };
 
 const runRequestToolAssistant = async ({
+  getSecret,
   callAI,
   callAIImage,
   getAssistant,
@@ -1044,9 +1064,13 @@ const runRequestToolAssistant = async ({
   sessionId,
   cb,
   toolParameter,
-  projectId,
   datastoreVariables,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   callAI: CallAI;
   callAIImage: CallAIImage;
   getAssistant: GetAssistant;
@@ -1056,7 +1080,6 @@ const runRequestToolAssistant = async ({
   user?: User;
   sessionId?: string;
   toolParameter: Parameter;
-  projectId?: string;
   datastoreVariables: Variable[];
 }) => {
   if (
@@ -1068,7 +1091,7 @@ const runRequestToolAssistant = async ({
     const currentTaskId = taskIdGenerator.nextId().toString();
 
     const { agent: tool } = toolParameter.source;
-    const toolAssistant = await getAssistant(tool.id);
+    const toolAssistant = await getAssistant(tool.id, { blockletDid: tool.blockletDid, projectId: tool.projectId });
     if (!toolAssistant) return null;
 
     const args = Object.fromEntries(
@@ -1084,6 +1107,7 @@ const runRequestToolAssistant = async ({
     );
 
     const result = await runAssistant({
+      getSecret,
       taskId: currentTaskId,
       callAI,
       callAIImage,
@@ -1094,7 +1118,7 @@ const runRequestToolAssistant = async ({
       callback: cb?.(currentTaskId),
       user,
       sessionId,
-      projectId,
+      projectId: tool.projectId!,
       datastoreVariables,
     });
 
@@ -1105,6 +1129,7 @@ const runRequestToolAssistant = async ({
 };
 
 const getVariables = async ({
+  getSecret,
   callAI,
   callAIImage,
   getAssistant,
@@ -1117,6 +1142,11 @@ const getVariables = async ({
   projectId,
   datastoreVariables,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   callAI: CallAI;
   callAIImage: CallAIImage;
   getAssistant: GetAssistant;
@@ -1126,7 +1156,7 @@ const getVariables = async ({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
 }) => {
   const variables: { [key: string]: any } = { ...parameters };
@@ -1145,13 +1175,22 @@ const getVariables = async ({
 
   for (const parameter of assistant.parameters || []) {
     if (parameter.key && parameter.type === 'source') {
-      if (parameter.source?.variableFrom === 'tool' && parameter.source.agent) {
+      if (parameter.source?.variableFrom === 'secret') {
+        const { secret } = await getSecret({
+          targetProjectId: projectId,
+          targetAgentId: assistant.id,
+          targetInputKey: parameter.key!,
+        });
+        if (!secret) throw new Error(`Missing required agent secret ${parameter.key}`);
+        variables[parameter.key!] = secret;
+      } else if (parameter.source?.variableFrom === 'tool' && parameter.source.agent) {
         const { agent: tool } = parameter.source;
-        const toolAssistant = await getAssistant(tool.id);
+        const toolAssistant = await getAssistant(tool.id, { blockletDid: tool.blockletDid, projectId: tool.projectId });
         if (!toolAssistant) continue;
 
         // eslint-disable-next-line no-await-in-loop
         const result = await runRequestToolAssistant({
+          getSecret,
           callAI,
           callAIImage,
           getAssistant,
@@ -1161,7 +1200,6 @@ const getVariables = async ({
           sessionId,
           cb,
           toolParameter: parameter,
-          projectId,
           datastoreVariables,
         });
 
@@ -1288,6 +1326,7 @@ async function runAgent({
 }
 
 async function runPromptAssistant({
+  getSecret,
   callAI,
   callAIImage,
   taskId,
@@ -1302,6 +1341,11 @@ async function runPromptAssistant({
   datastoreVariables,
   functionName,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   callAI: CallAI;
   callAIImage: CallAIImage;
   taskId: string;
@@ -1312,7 +1356,7 @@ async function runPromptAssistant({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
   functionName?: string;
 }) {
@@ -1321,6 +1365,7 @@ async function runPromptAssistant({
     .map((i) => i.data);
 
   const blockResults = await runExecuteBlocks({
+    getSecret,
     assistant,
     callAI,
     callAIImage,
@@ -1598,6 +1643,7 @@ async function runPromptAssistant({
 }
 
 async function runImageAssistant({
+  getSecret,
   callAI,
   callAIImage,
   taskId,
@@ -1611,6 +1657,11 @@ async function runImageAssistant({
   projectId,
   datastoreVariables,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   callAI: CallAI;
   callAIImage: CallAIImage;
   taskId: string;
@@ -1621,13 +1672,14 @@ async function runImageAssistant({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
 }) {
   if (!assistant.prompt?.length) throw new Error('Prompt cannot be empty');
 
   const blockResults = assistant.prepareExecutes?.length
     ? await runExecuteBlocks({
+        getSecret,
         assistant,
         callAI,
         callAIImage,
@@ -1734,6 +1786,7 @@ async function runImageAssistant({
 }
 
 async function runExecuteBlocks({
+  getSecret,
   assistant,
   callAI,
   callAIImage,
@@ -1747,6 +1800,11 @@ async function runExecuteBlocks({
   projectId,
   datastoreVariables,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   assistant: Assistant;
   callAI: CallAI;
   callAIImage: CallAIImage;
@@ -1757,7 +1815,7 @@ async function runExecuteBlocks({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
 }) {
   if (!executeBlocks.length) return [];
@@ -1771,6 +1829,7 @@ async function runExecuteBlocks({
   for (const executeBlock of executeBlocks) {
     const task = () => async () => {
       cache[executeBlock.id] ??= runExecuteBlock({
+        getSecret,
         taskId: taskIdGenerator.nextId().toString(),
         assistant,
         callAI,
@@ -1801,6 +1860,7 @@ async function runExecuteBlocks({
 }
 
 async function runExecuteBlock({
+  getSecret,
   taskId,
   assistant,
   callAI,
@@ -1816,6 +1876,11 @@ async function runExecuteBlock({
   projectId,
   datastoreVariables,
 }: {
+  getSecret: (args: {
+    targetProjectId: string;
+    targetAgentId: string;
+    targetInputKey: string;
+  }) => Promise<{ secret: string }>;
   taskId: string;
   assistant: Assistant;
   callAI: CallAI;
@@ -1828,7 +1893,7 @@ async function runExecuteBlock({
   callback?: RunAssistantCallback;
   user?: User;
   sessionId?: string;
-  projectId?: string;
+  projectId: string;
   datastoreVariables: Variable[];
 }) {
   const { tools } = executeBlock;
@@ -1896,7 +1961,10 @@ async function runExecuteBlock({
             });
           }
 
-          const toolAssistant = await getAssistant(tool.id);
+          const toolAssistant = await getAssistant(tool.id, {
+            blockletDid: tool.blockletDid,
+            projectId: tool.projectId,
+          });
           if (!toolAssistant) return undefined;
           const args = Object.fromEntries(
             await Promise.all(
@@ -1912,6 +1980,7 @@ async function runExecuteBlock({
           );
 
           return runAssistant({
+            getSecret,
             taskId: currentTaskId,
             callAI,
             callAIImage,
@@ -1999,7 +2068,10 @@ async function runExecuteBlock({
             };
           }
 
-          const toolAssistant = await getAssistant(tool.id);
+          const toolAssistant = await getAssistant(tool.id, {
+            blockletDid: tool.blockletDid,
+            projectId: tool.projectId,
+          });
           if (!toolAssistant) return undefined;
           const toolParameters = (toolAssistant.parameters ?? [])
             .filter((i): i is typeof i & Required<Pick<typeof i, 'key'>> => !!i.key && !tool.parameters?.[i.key])
@@ -2171,6 +2243,7 @@ async function runExecuteBlock({
           );
 
           const res = await runAssistant({
+            getSecret,
             taskId: currentTaskId,
             callAI,
             callAIImage,